// frontend/src/api/model/githubCom4H1RZooraInternalDomainInvoice.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainInvoiceItem } from './githubCom4H1RZooraInternalDomainInvoiceItem';
import type { GithubCom4H1RZooraInternalDomainInvoiceStatus } from './githubCom4H1RZooraInternalDomainInvoiceStatus';
import type { GithubCom4H1RZooraInternalDomainPayment } from './githubCom4H1RZooraInternalDomainPayment';

export interface GithubCom4H1RZooraInternalDomainInvoice {
  created_at?: string;
  /** admin; nil = self-serve */
  created_by?: string;
  currency?: string;
  description?: string;
  /** pending payment deadline */
  expires_at?: string;
  id?: string;
  issued_at?: string;
  items?: GithubCom4H1RZooraInternalDomainInvoiceItem[];
  /** assigned at issue */
  number?: string;
  organization_id?: string;
  paid_at?: string;
  payments?: GithubCom4H1RZooraInternalDomainPayment[];
  status?: GithubCom4H1RZooraInternalDomainInvoiceStatus;
  subtotal?: number;
  tax_amount?: number;
  tax_percent?: number;
  total?: number;
  updated_at?: string;
}