// frontend/src/api/model/githubCom4H1RZooraInternalDomainAdminCreateInvoiceItemDTOKind.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainAdminCreateInvoiceItemDTOKind = typeof GithubCom4H1RZooraInternalDomainAdminCreateInvoiceItemDTOKind[keyof typeof GithubCom4H1RZooraInternalDomainAdminCreateInvoiceItemDTOKind];


export const GithubCom4H1RZooraInternalDomainAdminCreateInvoiceItemDTOKind = {
  plan_subscription: 'plan_subscription',
  custom: 'custom',
  addon: 'addon',
} as const;